Add author_details page
New page similar to repo details. The two pages are almost identical,
but a lot of code had to be duplicated to accommodate the author
specific statistics.

Also fix artifacts in graphs caused by date ordering. If the array
given to plotly is not sorted by date then the graph will have blatant
mistakes. Add a call to order_by when filtering to fix this