Added date picker feature
All the code for the date picker is in the header.html template for now
Modified the nav.js template by replacing all instances of "&end_str="
with "&end=" in order to keep the js regex that replaces the dates in the
query string as basic as possible. I also think that it helps with
the standardization of the query string.